requirements.txt
django
django-crispy-forms
django-cleanup
pillow
django-ckeditor